Refactor : 불필요한 컴포넌트 분리 개선
